requirements.txt
asgiref==3.11.0
Django==5.2.9
django-cors-headers==4.9.0
djangorestframework==3.16.1
sqlparse==0.5.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
dj-database-url==2.3.0
reportlab==4.4.6
openai==2.9.0
